server/requirements.txt
alembic==1.11.1
annotated-types==0.5.0
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.2
botocore==1.31.2
certifi==2023.5.7
click==8.1.4
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
Mako==1.2.4
MarkupSafe==2.1.3
mistune==3.0.1
packaging==23.1
psycopg2==2.9.6
pydantic==2.0.2
pydantic_core==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
referencing==0.29.1
rpds-py==0.8.10
s3transfer==0.6.1
sentry-sdk==1.28.0
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==1.26.16
voluptuous==0.13.1
Werkzeug==2.3.6
WTForms==3.0.1
